move functions meant for testing into test files
